Move reducers to their own section
