+++ src/map-interface/app-state/reducers/menu.ts
